+++ src/main/java/com/nexmo/client/account/PrefixPricingResponse.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2018 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,21 +19,33 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.verify;
+package com.nexmo.client.account;
 
-public class VerifyResult extends BaseResult {
-    private final String requestId;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nexmo.client.NexmoUnexpectedException;
 
-    public VerifyResult(final int status,
-                        final String requestId,
-                        final String errorText,
-                        final boolean temporaryError) {
-        super(status, errorText, temporaryError);
-        this.requestId = requestId;
+import java.io.IOException;
+import java.util.List;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class PrefixPricingResponse {
+    private int count;
+    private List<PricingResponse> countries;
+
+    public int getCount() {
+        return count;
     }
 
-    public String getRequestId() {
-        return this.requestId;
+    public List<PricingResponse> getCountries() {
+        return countries;
     }
 
+    public static PrefixPricingResponse fromJson(String json) {
+        try {
+            return new ObjectMapper().readValue(json, PrefixPricingResponse.class);
+        } catch (IOException jpe) {
+            throw new NexmoUnexpectedException("Failed to produce PrefixPricingResponse from json.", jpe);
+        }
+    }
 }