+++ src/main/java/com/nexmo/client/account/PricingEndpoint.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2018 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,21 +19,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.verify;
+package com.nexmo.client.account;
 
-public class VerifyResult extends BaseResult {
-    private final String requestId;
+import com.nexmo.client.HttpWrapper;
+import com.nexmo.client.NexmoClientException;
 
-    public VerifyResult(final int status,
-                        final String requestId,
-                        final String errorText,
-                        final boolean temporaryError) {
-        super(status, errorText, temporaryError);
-        this.requestId = requestId;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+
+class PricingEndpoint {
+    private Map<ServiceType, PricingMethod> methods = new HashMap<>();
+
+    PricingEndpoint(HttpWrapper httpWrapper) {
+        this.methods.put(ServiceType.SMS, new SmsPricingMethod(httpWrapper));
+        this.methods.put(ServiceType.VOICE, new VoicePricingMethod(httpWrapper));
     }
 
-    public String getRequestId() {
-        return this.requestId;
-    }
+    PricingResponse getPrice(ServiceType serviceType, PricingRequest request) throws IOException, NexmoClientException {
+        if (this.methods.containsKey(serviceType)) {
+            return this.methods.get(serviceType).execute(request);
+        }
 
+        throw new IllegalArgumentException("Unknown Service Type: " + serviceType);
+    }
 }