+++ src/main/java/com/nexmo/client/sms/SmsSingleSearchResponse.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2018 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,21 +19,22 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.verify;
+package com.nexmo.client.sms;
 
-public class VerifyResult extends BaseResult {
-    private final String requestId;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nexmo.client.NexmoUnexpectedException;
 
-    public VerifyResult(final int status,
-                        final String requestId,
-                        final String errorText,
-                        final boolean temporaryError) {
-        super(status, errorText, temporaryError);
-        this.requestId = requestId;
+import java.io.IOException;
+import java.text.SimpleDateFormat;
+
+public class SmsSingleSearchResponse extends SmsDetails {
+    public static SmsSingleSearchResponse fromJson(String json) {
+        try {
+            ObjectMapper mapper = new ObjectMapper();
+            mapper.setDateFormat(new SimpleDateFormat("yyyy-MM-dd HH:mm:ss"));
+            return mapper.readValue(json, SmsSingleSearchResponse.class);
+        } catch (IOException jpe) {
+            throw new NexmoUnexpectedException("Failed to produce SmsSingleSearchResponse from json.", jpe);
+        }
     }
-
-    public String getRequestId() {
-        return this.requestId;
-    }
-
 }