+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.6.0] - 2018-07-06
+### Added
+- Added `getSmsPrice` to `AccountClient` for getting SMS pricing for a country.
+- Added `getVoicePrice` to `AccountClient` for getting voice pricing for a country.
+- Added `getPrefixPrice` to `AccountClient` for getting SMS and voice pricing for a prefix.
+- Added `topUp` to `AccountClient` for topping up your account which has auto-reload enabled.
+- Added `getSms` to `SmsClient` for searching for a single message by id.
+- Added `ConversionClient` and the ability to interact with the Nexmo Conversion API.
+
 ## [3.5.0] - 2018-05-29
 ### Changed
 - Updated `VerifyClient` to use the JSON endpoints instead of XML.